add displayBook function
This commit adds a new function displayBook that loops over the myLibrary array and creates a new card element for each book. It also adds each property of the book object as a paragraph element to the card using insertAdjacentElement(). The card element is then appended to the DOM. This function makes it easier to render the books in the library and display them on the page.